requirements.txt
openai
secure-smtplib
requests
python-dotenv
fastapi
sqlalchemy
